+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Master
 
 * Improvements to `danger init` - orta
+* Underlying Danger under-the-hood changes to prepare for supporting JS Dangerfiles - orta
+  - If you see any weirdness around the plugin attributes table, or verbose output, it's from this. 
 
 ## 2.1.1